#RO-12 Regist Company year off Api - change attributes of CompanyYearOffRegistRequest to String - edit some code in RequestServiceImpl

+++ RequestOFF/src/main/java/com/dactech/requestoff/service/impl/RequestServiceImpl.java
@@ -50,7 +50,7 @@ public class RequestServiceImpl implements RequestService{
 			request.setValidFlag(Integer.parseInt(requestRegistRequest.getValidFlag()));
 		}
 		else {	//update request
-			long id = Integer.parseInt(requestRegistRequest.getId());
+			long id = Long.parseLong(requestRegistRequest.getId());
 			request = requestRepository.findById(id);
 			if(request == null) {
 				throw new Exception("Can not find request with id = " + id);
